✨ feat: enhance logging functionality and improve socket handling
src/stores/bot/logs.ts: Add $reset function to reset log state variables.
src/views/logsbot/components/LogsCardView.vue: Update log handling to include remaining logs and improve socket connection management.
