+++ src/stores/bot/logs.ts
@@ -11,7 +11,21 @@ export const useStoreLogsBot = defineStore("logs", () => {
   const status = ref<string>("Iniciado");
   const current_pid = ref("");
   const contentRef = ref<HTMLDivElement | null>(null);
+
+  function $reset() {
+    listLogs.value = [];
+    totalLogs.value = 0;
+    remainingLogs.value = 0;
+    currentPos.value = 0;
+    totalSuccess.value = 0;
+    totalErrors.value = 0;
+    status.value = "CARREGANDO";
+    current_pid.value = "";
+    contentRef.value = null;
+  }
+
   return {
+    $reset,
     listLogs,
     totalLogs,
     remainingLogs,
